[수정] 각 게시글 별 app 분리
