+++ .gitignore
@@ -21,6 +21,10 @@ media
 *.py[cod]
 *$py.class
 
+**/migrations/**
+!**/migrations
+!**/migrations/__init__.py
+
 # C extensions
 *.so
