+++ bestchoice/settings.py
@@ -48,7 +48,10 @@ INSTALLED_APPS = [
     'rest_framework',
     'rest_framework_simplejwt',
     'users',
-    'articles',
+    'festivals',
+    'joins',
+    'reviews',
+    'recruits',
 ]
 
 MIDDLEWARE = [
