+++ bestchoice/urls.py
@@ -21,7 +21,10 @@ from django.conf.urls.static import static
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('users/', include('users.urls')),
-    path('articles/', include('articles.urls')),
+    path('festivals/', include('festivals.urls')),
+    path('joins/', include('joins.urls')),
+    path('reviews/', include('reviews.urls')),
+    path('recruits/', include('recruits.urls')),
 ]
 
 urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
