+++ reviews/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ArticlesConfig(AppConfig):
+class ReviewsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "articles"
+    name = "reviews"
